Materials editor

+++ Textures/Texture.h
@@ -1,7 +1,8 @@
 #pragma once
 #include "Windows.h"
-#include <d3d12.h>
+
 #include <Types.h>
+#include <d3d12.h>
 #include <wrl/client.h>
 
 struct STexture
@@ -12,5 +13,5 @@ struct STexture
 	Microsoft::WRL::ComPtr<ID3D12Resource> Resource = nullptr;
 	Microsoft::WRL::ComPtr<ID3D12Resource> UploadHeap = nullptr;
 
-	int32_t HeapIdx = -1;
+	int64_t HeapIdx = -1;
 };
